move getLabel fct to parent class
